Add handling of HTTP Authorization header to test-server.py
I think this has been done to HTTP specs.  It appears that
the Authorization header gets converted into AUTH_TYPE
and HTTP_AUTHORIZATION environment variables for the CGI
scripts.  However, I have also read online that Apache
does not pass these variables to CGI scripts by default,
for security reasons.  So this may require using a different
mechanism in the CGI script when it is used with "real"
web servers.

Signed-off-by: Tim Bird <tim.bird@sony.com>