Refactor some object management routines
Add sone "_api_" versions of get_object_map and get_object_data,
which return responses in "api" format (json data with the result
in the 'result' field, error message in the "message" field, etc.)

Convert some of the "non-api" routines to log errors intead of
return them to the user.  Probably, these should saved in an error
stack in the request, and it should be decided later whether to
return them as part of the html response or not.

Signed-off-by: Tim Bird <tim.bird@sony.com>