Remove lots of legacy code and data
The 'lc' command was originally copied from 'ftc' from the Fuego
project.  It had code for lots of items that were implemented
in a way differently for fserver than for the labcontrol.
There was too much junk lying around in the code, and it
made the code difficult to maintain.  So remove it.

This cuts the code roughly in half.

Some of the features may yet make their way back (like
support for request objects, the setenv command, or the
ability to query objects by attribute on the server side)
However, move these into an lc.legacy file, and only import
things as they are put into production.

Signed-off-by: Tim Bird <tim.bird@sony.com>