Support config in user's home dir
Add support for checking for ~/.lc.conf as the primary
configuration.  If that's not found, fall back to the
system configuration in /etc/lc.conf.  Either of these
are overridden by using the '-c <conf_file' option.

Also, remove a few more legacy items (unused imports
and more where help.)

Signed-off-by: Tim Bird <tim.bird@sony.com>