Avoid json-ifying data from serial captures
There's a temporary hardcoded data translation in the get_data
routine, for power management data.  It is converted from a CSV input
format to an json array of timestamped values.  Remove this translation
for serial capture data.

Also, add support for saving updated config information back out to the
resource file.  This is needed because grabserial needs to set the baud
rate, based on the desired configuration.  grabserial should really have
a way to use the existing baud rate configured for the port, but for
now, keep track of the baud_rate in the resource json file, and use it
on the grabserial command line.

Also, do some other code cleanups: finish a rename from 'action' to
'res_type', and restructure code for handling the success case for
start_capture.