Add initial (partial) support for 'lc upload'
Add the initial support for the upload operation.  The server part is
not quite finished, and does not parse the formatted multi-part form
data as formatted by the python requests module.
However, I wanted to get other commits pushed to the master branch.
Hence this partial commit.  This code might change, and the feature
should be considered unstable.

Signed-off-by: Tim Bird <tim.bird@sony.com>