Change list response data to include result field
Change the list reponse to have result field, and put
the list data in the 'data' field.  I think TimeSys changed
these APIs (list devices and mydevices) so that they had
the same format as other responses from the server.

That's good.  Change lcserver and lc to match.

Change the server to send a structured response (not
just the list array), and change the client to receive
and interpret the structured response.

Signed-off-by: Tim Bird <tim.bird@sony.com>