Refactor return_api_board_action (and add upload)
Refactor return_api_board_action to put handling of individual
operations on a board into sub-routines.  The routine was getting
too long to add more operations to.   This was in preparation for
adding support for the 'upload' function.

Also, the first version (alpha) of the code to support
upload operations was added.

Signed-off-by: Tim Bird <tim.bird@sony.com>