Add support for showing pages
Add page_data_class, to allow substituting data elements into a
page that is read from disk.  Pages live in lc-data/pages.
If the page_name is not a recognized page, then check that directory
to see if it matches a filename.  If so, read the data from the file
and interpolate data items and screen elements onto it.

Create a Test_Page_Data page and a Main page, to start with.
Refactor the do_show() routine to handle showing pages.

Also, turn on debugging if the file 'debug' exists in the
lc-data directory.  This allows for turning script debugging on
or off without modifyin the script.

Signed-off-by: Tim Bird <tim.bird@sony.com>