Handle both auth and cookie tokens in set_user
Allow set_user to handle authentication tokens that are
stored in cookies.  This allows users at the user interface
to be authenticated and perform restricted activities, once
they have logged in.

Return any cookies in the HTML response header.  The current
structure requires that cookies are set before calling req.show_header()
So far, this is only done by login and logout actions, so refactor
those to not emit html until after the login or logout operations
are performed (and the cookies are set appropriately).

Signed-off-by: Tim Bird <tim.bird@sony.com>