Remove HTTP_AUTH-related variables from env, when not present
Try to fix an issue with the server continuing to set some
environment variables, even when they are not on the request.
Since the server is long-lived, it saves these environment
variables and they get passed to the CGI script, even when
they are not set in the HTTP request.

This results in a logged out user getting logged in when
the lc client is run.  The logic here needs to get fixed.

Signed-off-by: Tim Bird <tim.bird@sony.com>