Change the sample lcserver.conf files_url_base setting
This points to the top of the data directory, not internally
into the files.  Right now this is a security issue, since
it means that someone can read the json files directly for
user, resource, boards, etc.  For users, this exposes their
passwords, which is not good.  Make a private directory
and migrate security-sensitive stuff there, as the labcontrol
server developers.  For now, this is OK.

Signed-off-by: Tim Bird <tim.bird@sony.com>