lcserver: parse run command from form, if not in json
A standard web form sends the data to the server as regular form
data (not in json), so if parsing the json data (from 'lc') doesn't
work, try parsing it directly from the form.

Handle the case where the command to run could not be parsed, by
returning an error.

Signed-off-by: Tim Bird <tim.bird@sony.com>