lcserver: use timestamped filename for last image and video
Use the timestamped filename for the last image and video file
when displaying these on the board information page.

This avoids having the browser cache the copy of image from the
"...last-camera-image.jpeg" file, which is wrong when the symlink
target is updated (ie, when a new picture is taken).

Signed-off-by: Tim Bird <tim.bird@sony.com>