lcserver: Add support for timed reservations
Add 'default_reservation_duration' to the server config.
Parse the duration, in minutes, from the API url.
Set the start_time and end_time for the reservation in the board file,
on allocation/reservation and release.
Show ending time on error messages from 'assign' operation.

Also, add support for showing reservations in the users page.

Note that the expiration of the allocation is not currently handled,
yet.

This is based on a patch submitted by Pavan Arun Deshpande, with
significant modifications in some places.

Signed-off-by: Tim Bird <tim.bird@sony.com>