lcserver: expire timed registrations, whenever the board data is read
Add a check_reservation_expiry() routine, and call it from several
locations.  Probably the calls from do_board_assign and do_board_release
are not needed, as the calls in get_object_map() and get_api_object_map()
will cover every case I can think of.  This means we check the status
quite often.

This causes quite a few more calls to get the current datetime, but
shouldn't result in any additional file or json parsing operations,
unless the reservation expired.

For a user at the web interface, or using the command line, this has
the effect of appearing that the server expired this when it should
have (but the data will be sitting on disk, unexpired, until the
server changes it on the next access.)

It seems to work.

Signed-off-by: Tim Bird <tim.bird@sony.com>