lcserver: support camera operations
In order to return full urls, add url_prefix to the server
configuration.  Change do_camera_operation to do_board_camera_operation
to make this consistent with other functions that handle actions
relative to a board instead of a resource.  Return just the
straight url_path for a still image after a camera operation
is performed.

Change start_command to look for programs in the labcontrol/utils
directory.

rename action to res_type for several of sub-routines. Parse duration
for a camera resource type in start_capture(). Add logpath to the
pidfile, for easier access later.  Add 'capture()' function to retrieve
a single still image (and report its url_path).

Add get_captured_data_ref() to handle "get-ref" operation.

And (finally) handle multiplexing for the camera APIs.

This is not finished yet, but it works for still images for now.

Signed-off-by: Tim Bird <tim.bird@sony.com>