lc: detect json in download return data
In the error case, a download operation will return json, which
can be parsed to get the error reason.  Do this so that I can
print a better message than from the failure later in the
file: "Error: Cannot perform tar for 'download' operation"
which is unhelpful when the problem is a server-side issue
with the download_cmd definition in the resource file.

Signed-off-by: Tim Bird <tim.bird@sony.com>