lcserver: first use of javascript to show uptime on show_board page
This is an experimental use of javascript attached to a button,
to show running a command on the server and displaying the results
on the current page.

I'd like to convert most of our buttons to this style of action.

This is just prototype code.  The real uptime will likely be
timer-based.

Signed-off-by: Tim Bird <tim.bird@sony.com>