Change wire protocol for set-config
Instead of sending json to the server for the set-config operation,
send x-www-form-urlencoded (HTTP FORM) data instead.  This is
accomplished by converting the json input (from stdin) into a
python dictionary, and handing that to the request module on
the client side.  On the server side, parse the form data
directly from the form parsed by the python cgi module.

The old code (parsing the raw body of the POST message as json)
worked with the test web server, but it doesn't appear to work
with Apache.  I'm not sure why not, or what changed.
But the req.form.value just returns an empty list, instead
of the json string.

Signed-off-by: Tim Bird <tim.bird@sony.com>