Add support for a prototype 'run2' operation
The 'run2' operation is just a prototype, but it adds support
for splitting the stream of the executed program into stdout
and stderr, by embedded marker delimiters in the single
response stream return by the server.  It works, and needs more
testing.  Curl and sh will not be able to handle this, so I'm not
sure if the ebf client can support this.  But this is for
proof-of-concept.

Signed-off-by: Tim Bird <tim.bird@sony.com>