lcserver: add support for reporting stderr of capture command
If the capture command has some problem after it starts,
there's no way to report it from start_capture().  So save
the stderr of the program to a file, and on stop_capture,
check the file and send that to the client as an error message.

This should help diagnose problems with the capture command.
This was inspired by a problem with grabserial, where the
wrong 'serial' module was installed, and I was getting an
exception.  The exception was going to stderr, but the server
was just discarding that, and it was very hard to debug.

Now, the stderr from the capture command is both 1) sent
back to the client, and 2) saved to the log.

Signed-off-by: Tim Bird <tim.bird@sony.com>