Add acme-capture.sh for capturing power data from an ACME board
Implement capture of power data from a board, retrieving data
from the sysfs filesystem entries for the board (as an i2c device)
using a custom script and ssh.

Place the data in a logfile as specified on the command line.
Correctly handle SIGTERM, to stop the capture and clean up items
on the board.

Signed-off-by: Tim Bird <tim.bird@sony.com>