lcserver: disable conversion of power-measurement data
Previously, during testing with the Sony debug board
resources, I converted the data from the board into an
array of lines, in json format.  The ACME board just
returns CSV, and that's what I think we're going to
standardize on, for PM data, for now.

Disable this conversion (but leave the code there for
future reference), by making the string wrong in the
conditional which checks whether to perform this conversion

Signed-off-by: Tim Bird <tim.bird@sony.com>