lcserver: Add navbar, and refactor pages and views
Add CSS for the navbar, and make the login stuff float right.
Refactor req.show_header into req.show_html_header and
a global show_header() function, to show the navbar on every page.
Remove the old trailing bullet menu from do_show()
Move the raw object view to the Admin page.

This is a general overhaul of the UI to be more friendly and compact.

Signed-off-by: Tim Bird <tim.bird@sony.com>