Change start_server to use python3
The rest of the project is upgrading to python3.
Have the non-production test server use that also.
This means that test-server.py is no longer needed,
but I'm not going to rip it out just yet.

Save the old python2-based start_server, in case
that it useful for something.  But rename it so that
it's use requires specific intention to invoke it.

Signed-off-by: Tim Bird <tim.bird@sony.com>