Add some example live-stream start and stop scripts
Add some scripts to start and stop live-streaming to the utils
directory.  These are different from other utility scripts,
because they are intended to be run on the resource itself,
not on the host where labcontrol is running.

The original (first) script (live_stream.sh) was written by
Pavan Arun Deshpande, and others are derivative of that.

Signed-off-by: Tim Bird <tim.bird@sony.com>