Add stream-proxy - first draft (rough)
Add a new CGI script for doing web proxying of live streamining
services.  I don't even know if this can be made to work, but
this one uses a data file to redirect a request to a stream-proxy
URL to a lab-local live streaming URL.

This is very much still a work-in-progress.

Signed-off-by: Tim Bird <tim.bird@sony.com>