lcserver: Add support for DUT web terminal
Use ttyd as the terminal provider.  Add plumbing and process lifecycle
management to lcserver, to start ttyd when requested, and provide an
iframe for the web terminal itself (which is delivered from ttyd
as a javascript xterm console).

Add routines to read and write process data.  Calculate an unused port
for ttyd for a new web terminal server instance.  Add a link to
show_board_info() for the terminal page, when a board has 'login_cmd'.

To activate this feature for a board, add a 'login_cmd' in the
board file for a board.  It can be something like this:
    sshpass -p %(password)s ssh -q -o UserKnownHostsFile=/dev/null -o "StrictHostKeyChecking no" -x %(user)s@%(ip_addr)s

The board file should have password, user and ip_addr defined.

There are some remaining issues, such as terminating the ttyd server
instance, and checking the board reservation before allowing access
to the web terminal. But this works well enough to test and refine now.

Signed-off-by: Tim Bird <tim.bird@sony.com>