lcserver: change URL for webterm
Use a path element instead of the socket, for the proxied webterm
connection.  See the NOTES-DUT-xterm-support.txt file for information
about how to set up Apache so that the webterm connection is
proxied directly by the same Apache server where lcserver.py runs.
This means that the client does NOT have to connect directly to
another port on the host for this connection.

Signed-off-by: Tim Bird <tim.bird@sony.com>