server: Use incremental backoff for stop_capture sigterm
When terminating the capture process during a stop_capture
operation, wait longer than the current 2 seconds.  Some
audio captures take a long time to terminate.  Use an
increasing incremental backoff for the delays between
sending the SIGTERM signal to the capture process.
This still retains relatively good performance when the
capture process terminates quickly, while still allowing
for longer timeouts and not a lot of overhead, for processes
that take longer to terminate.

Signed-off-by: Tim Bird <tim.bird@sony.com>