utils: add a utility for executing arecord on the labcontrol host
Add alsa-record.sh to allow labcontrol to record audio on
the host where labcontrol is located.  This is used when a resource
for audio recording (a mic) is on the same host as the labcontrol
server.  If the audio input is on another device, a more elaborate
wrapper is needed.

Signed-off-by: Tim Bird <tim.bird@sony.com>