Add sample audio playback test
These tests are not finalized, but they demonstrate how to perform
and audio playback test using a DUT for playback and a lab resource
for capturing sound.  This eliminates the need for a loopback
device on the DUT.

There are some hardcoded elements to these scripts, that need to
be removed for production-grade testing.  However, these scripts
show basic techniques for audio testing.

Note that this uses a modified alsabat program for audio analysis
at the end of the test, as well as unmodified 'sox' for test
operation and data analysis.

Signed-off-by: Tim Bird <tim.bird@sony.com>